requirements.txt
uvicorn
fastapi
gunicorn
requests
python-decouple
groq
numpy==1.23.5
scipy==1.10.1
scikit-image==0.19.3
opencv-python==4.6.0.66
paddlepaddle==2.5.0
paddleocr==2.6.1.3
imgaug==0.4.0